requirements.txt
# Mobile Automation Dependencies
appium-python-client==3.1.1
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
selenium==4.15.2
allure-pytest==2.13.2
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0